requirements.txt
python-jenkins==1.8.1